feat: refactor some code to keep order
